Default dtype and device to SB environment

+++ tests/models/test_number_partitioning.py
@@ -6,7 +6,7 @@ from src.simulated_bifurcation.models import NumberPartitioning
 def test_number_partitioning_with_even_sum():
     torch.manual_seed(42)
     numbers = [2, 7, 3, 8, 1, 5, 4, 6]
-    model = NumberPartitioning(numbers)
+    model = NumberPartitioning(numbers, dtype=torch.float32)
     model.optimize(agents=10, verbose=False)
     result = model.partition
     assert result["left"]["sum"] == result["right"]["sum"]
@@ -15,7 +15,7 @@ def test_number_partitioning_with_even_sum():
 def test_number_partitioning_with_odd_sum():
     torch.manual_seed(42)
     numbers = [2, 7, 3, 9, 1, 5, 4, 6]
-    model = NumberPartitioning(numbers)
+    model = NumberPartitioning(numbers, dtype=torch.float32)
     model.optimize(agents=10, verbose=False)
     result = model.partition
     assert abs(result["left"]["sum"] - result["right"]["sum"]) == 1
@@ -24,7 +24,7 @@ def test_number_partitioning_with_odd_sum():
 def test_number_partitioning_with_gap():
     torch.manual_seed(42)
     numbers = [431, 444, 654, 949, 17, 707, 482, 179, 442, 949]
-    model = NumberPartitioning(numbers)
+    model = NumberPartitioning(numbers, dtype=torch.float32)
     model.optimize(agents=64, verbose=False)
     result = model.partition
     assert abs(result["left"]["sum"] - result["right"]["sum"]) == 10
